bloggin_system/requirements.txt
Django
djangorestframework
fastapi
uvicorn
djangorestframework-simplejwt
django-redis
psycopg2-binary
pyjwt
pydantic
django-cors-headers
python-multipart
python-dotenv